% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_wrapper.R
\name{adapt_mtp}
\alias{adapt_mtp}
\title{Wrapper of AdaPT multiple testing procedure}
\usage{
adapt_mtp(Ps, Xs, alpha, formula_rhs = "~X1+X2", return_fit = FALSE)
}
\arguments{
\item{Ps}{Numeric vector of p-values}

\item{Xs}{Data frame with features}

\item{alpha}{Nominal testing level}

\item{formula_rhs}{Formula defining the RHS in the fitted GLMs, defaults to ~X1+X2 (used in simulations herein).}

\item{return_fit}{Boolean, whether to return the fitted adapt object (or only the indicator of rejections), defaults to false}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Wrapper of AdaPT multiple testing procedure
}
\references{
AdaptMT CRAN package
}
