% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_simulations_functions.R
\name{apply_grouped_methods}
\alias{apply_grouped_methods}
\title{Apply grouped multiple testing methods to simulated grouped multiple testing data}
\usage{
apply_grouped_methods(sim, alpha, gbh_only = FALSE)
}
\arguments{
\item{sim}{Data frame with simulation data, e.g., as returned by the `grouped_global_null_sim` function.}

\item{alpha}{Nominal testing level}

\item{gbh_only}{Whether to only apply GBH and IHW-GBH}
}
\value{
Data frame with FDP and Power of different methods on this simulation.
}
\description{
Apply grouped multiple testing methods to simulated grouped multiple testing data
}
