% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamix_simulations_functions.R
\name{beta_unif_sim}
\alias{beta_unif_sim}
\title{Simulation: Misspecified conditional Beta-uniform mixture model}
\usage{
beta_unif_sim(
  m = 10000,
  mus_slope = 1.5,
  one_sided_tests = FALSE,
  prob_one_sided = 0.25
)
}
\arguments{
\item{m}{Number of hypotheses (default: m=10000)}

\item{mus_slope}{Numeric (default:1.5) parameter bar(beta) in equation (12)}

\item{one_sided_tests}{Bool (default:FALSE), if true adds some nulls that are strictly superuniform}

\item{prob_one_sided}{Numeric (default:0.25) proportion of nulls that are strictly superuniform}
}
\value{
Data frame with columns `Hs` (null or alternative), `Ps` (p-value), `Xs` (side-information),
        `alphas` (parameters of alternative distribution), `pi1s` (probabilities of being from the alternative distribution),
        `oracle_lfdr` (oracle local fdr)
}
\description{
Simulation: Misspecified conditional Beta-uniform mixture model
}
