% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamix.R
\name{betamix_oracle_lfdr}
\alias{betamix_oracle_lfdr}
\title{Oracle local false discovery rate procedure in the beta mixture model}
\usage{
betamix_oracle_lfdr(Ps, pi1s, mu_alphas, alpha)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{pi1s}{True probability of being an alternative in the beta mixture model}

\item{mu_alphas}{Parameter vector of alternative Beta(mu_alpha, 1) distribution of each test.}

\item{alpha}{Significance level at which to apply method}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Oracle local false discovery rate procedure in the beta mixture model
}
