% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars.R
\name{CARS_extended}
\alias{CARS_extended}
\title{Modified CARS function}
\usage{
CARS_extended(t_1, t_2, alpha, tau = 0.9, option = "regular")
}
\arguments{
\item{t_1}{Numeric vector of primary z-score statistics (converted to p-values via 2*(1-pnorm(abs(primary_stats)))).}

\item{t_2}{Numeric vector of secondary statistic}

\item{alpha}{Significance level at which to control FDR}

\item{tau}{Threshold for choosing interesting locations for density estimation (default 0.9)}

\item{option}{Currently only supports 'regular' (and not sparse from the CARS package)}

\item{variance}{Numeric matrix with variances for each test/group.}
}
\value{
A list that contains `de`, i.e., the vector with the significant tests and `cars_fun` which is a function that
        maps t_1,t_2 to the estimated local fdr.
}
\description{
Compared to the `CARS::CARS` function, there are three main differences:
1) this function takes input that has already been summarized.
2) it returns additional information (cf. below).
3) we assume that we have Gaussian test-statistics with known variance.
}
