% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars_simulations_functions.R
\name{CARS_preprocess}
\alias{CARS_preprocess}
\title{Data preprocessing for CARS to be used with p-value methods}
\usage{
CARS_preprocess(X, Y, variance)
}
\arguments{
\item{X, Y, variance}{X,Y matrices with data for two groups used for two-sample multiple testing and per-test per-sample variance matrix, as returned
by the `cars_ttest_data` function.}
}
\value{
Data.frame with columns `pvalue` and `t_2` (ancillary --independent under the null-- testing statistic)
}
\description{
Data preprocessing for CARS to be used with p-value methods
}
\examples{
cars_example_data <- cars_ttest_data(m=10000, k=2000, n_x=50, n_y=50, mu_x_1 = 0.5, mu_x_2 = 0.25, mu_y_2 = 0.25)
cars_example_preprocess <- with(cars_example_data, CARS_preprocess(x, y, var_mat))
}
