% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{cars_weighter}
\alias{cars_weighter}
\title{Heuristic multiple testing weights based on CARS procedure}
\usage{
cars_weighter(primary_stats, Xs, Xs_new, tau, alpha)
}
\arguments{
\item{primary_stats}{Numeric vector of z-score statistics (converted to p-values via 2*(1-pnorm(abs(primary_stats)))).}

\item{Xs}{Numeric vector or matrix with covariates}

\item{Xs_new}{Numeric vector or matrix with covariates  (out-of-fold).}

\item{tau}{Numeric, the level at which tau-censoring is applied.}

\item{alpha}{Significance level at which to apply method (not used for this weight learning method)}
}
\value{
Numeric vector with weights.
}
\description{
Heuristic multiple testing weights based on CARS procedure
}
