% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{censored_betamix_weighter}
\alias{censored_betamix_weighter}
\title{Learn multiple testing weights based on EM fitting of the betamix model and local FDRs}
\usage{
censored_betamix_weighter(
  Ps,
  Xs,
  Xs_new,
  tau,
  alpha,
  formula_rhs = "~X1+X2",
  pi1_min = 0.01,
  pi1_max = 0.9,
  alpha_max = 0.9,
  maxiter = 200,
  numerator_bh = TRUE,
  ...
)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values in training set.}

\item{Xs}{Factor of groups in training set.}

\item{Xs_new}{Factor of groups in test-set (out-of-fold).}

\item{tau}{Numeric (default = 0.5), the level at which tau-censoring is applied.}

\item{alpha}{Significance level at which to apply method (not used for this weight learning method)}

\item{formula_rhs}{Formula defining the RHS in the fitted GLMs, defaults to ~X1+X2 (used in simulations herein).}

\item{pi1_min}{Numeric in (0,1),  lower bound on conditionally probability of being an alternative (default: 0.01).}

\item{pi1_max}{Numeric in (0,1), upper bound on conditionally probability of being a null (default: 0.9).}

\item{alpha_max}{Numeric in (0,1), upper bound on parameter of alternative Beta distribution (default: 0.9).}

\item{maxiter}{(Default: 200) Total number of iterations to run the EM algorithm}

\item{numerator_bh}{Bool (default:TRUE) Define the marginal FDR as sum pi_0(X_i) t(X_i), if FALSE, or sum t(X_i) if TRUE}
}
\value{
Numeric vector with weights of test-set hypotheses
}
\description{
Learn multiple testing weights based on EM fitting of the betamix model and local FDRs
}
