% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamix_simulations_functions.R
\name{eval_beta_unif_sim}
\alias{eval_beta_unif_sim}
\title{Apply multiple testing methods to the simulation with the continuous covariate.}
\usage{
eval_beta_unif_sim(
  mu_slope,
  seed,
  one_sided_tests = FALSE,
  prob_one_sided = 0.25,
  alpha = 0.1,
  m = 10000,
  lfdr_only = FALSE
)
}
\arguments{
\item{mu_slope}{mus_slope Numeric (default:1.5) parameter bar(beta) in equation (12)}

\item{seed}{Integer; used for printing which simulation it running (does not set an actual RNG seed)}

\item{one_sided_tests}{Bool (default:FALSE), if true adds some nulls that are strictly superuniform}

\item{prob_one_sided}{Numeric (default:0.25) proportion of nulls that are strictly superuniform}

\item{alpha}{Numeric (default: 0.1), nominal significance level at which to apply methods}

\item{m}{Number of hypotheses (default: m=10000)}

\item{lfdr_only}{Bool (default:FALSE), whether to run all methods (if FALSE) or only lfdr based methods (if TRUE)}
}
\value{
Data frame with FDP and Power of different methods on this simulation.
}
\description{
Apply multiple testing methods to the simulation with the continuous covariate.
}
