% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars_simulations_functions.R
\name{eval_cars_sim}
\alias{eval_cars_sim}
\title{Apply multiple testing methods to the simultaneous two-sample testing simulation.}
\usage{
eval_cars_sim(m, k, mu_x_1, mu_x_2, seed, n_x = 50, n_y = 50, alpha = 0.1)
}
\arguments{
\item{m}{Number of hypotheses (default: m=10000)}

\item{k}{Number of alternatives (default: k=10)}

\item{mu_x_1}{Signal strength for first locations of X}

\item{mu_x_2}{Signal strength for locations (k+1) to 2k of X}

\item{seed}{Integer; used for printing which simulation it running (does not set an actual RNG seed)}

\item{n_x}{Number of samples for X (first group) in each test (default: n_x=50)}

\item{n_y}{Number of samples for Y (second group) in each test (default: n_y=50)}

\item{alpha}{Numeric (default: 0.1), nominal significance level at which to apply methods}
}
\value{
Data frame with FDP and Power of different methods on this simulation.
}
\description{
Apply multiple testing methods to the simultaneous two-sample testing simulation.
}
