% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{fdp_eval}
\alias{fdp_eval}
\title{Evaluate multiple testing procedure}
\usage{
fdp_eval(Hs, rjs)
}
\arguments{
\item{Hs}{Vector with indicators of alternatives (1) and true nulls (0)}

\item{rjs}{Vector with indicator of rejected hypotheses}
}
\value{
Data frame with columns `rjs` (total rejections), `pow` (Power), `FDP` (False discovery proportion)
}
\description{
Evaluate multiple testing procedure
}
