% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamix.R
\name{gamma_glm_censored_em}
\alias{gamma_glm_censored_em}
\title{Fit conditional beta-uniform mixture model using EM algorithm under p-value tau-censoring.}
\usage{
gamma_glm_censored_em(
  censored_Ps,
  Xs,
  tau_censor,
  formula_rhs = "~X1+X2",
  maxiter = 50,
  tau_pi0 = 0.5,
  alpha_min = 0.1,
  pi1_min = 0.01,
  pi1_max = 0.9,
  alpha_max = 0.9
)
}
\arguments{
\item{censored_Ps}{Numeric vector of tau-censored p-values}

\item{Xs}{Data frame with features}

\item{tau_censor}{Numeric, level at which p-valeus have been tau-censored}

\item{formula_rhs}{Formula defining the RHS in the fitted GLMs, defaults to ~X1+X2 (used in simulations herein).}

\item{maxiter}{Total number of iterations to run the EM algorithm.}

\item{tau_pi0}{Number in (0,1) used for the initialization of the EM algorithm through the Boca-Leek approach (default: 0.5).}

\item{alpha_min}{Numeric in (0,1), upper bound on parameter of alternative Beta distribution (default: 0.1).}

\item{pi1_min}{Numeric in (0,1),  lower bound on conditionally probability of being an alternative (default: 0.01).}

\item{pi1_max}{Numeric in (0,1), upper bound on conditionally probability of being a null (default: 0.9).}

\item{alpha_max}{Numeric in (0,1), upper bound on parameter of alternative Beta distribution (default: 0.9).}
}
\value{
List with parameters of fitted conditional Beta-Uniform mixture
}
\description{
Fit conditional beta-uniform mixture model using EM algorithm under p-value tau-censoring.
}
