% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_methods.R
\name{gbh_simple}
\alias{gbh_simple}
\title{Group Benjamini Hochberg (GBH) multiple testing procedure}
\usage{
gbh_simple(unadj_p, groups, alpha, tau = 0.5, Storey = FALSE)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{groups}{Factor to which different hypotheses belong}

\item{alpha}{Significance level at which to apply method}

\item{tau}{Numeric (default: 0.5) level at which to apply Storey's pi0 estimator}

\item{Storey}{Bool (default: FALSE): is the GBH procedure pi0 adaptive or not?}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Group Benjamini Hochberg (GBH) multiple testing procedure
}
