% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamix.R
\name{get_tailfdr_betamix}
\alias{get_tailfdr_betamix}
\title{Tail fdr for beta-mixture model}
\usage{
get_tailfdr_betamix(ts, pi1s, alphas, numerator_bh = FALSE)
}
\arguments{
\item{ts}{Numeric vector of thresholds at which to compute the local fdr}

\item{pi1s}{Numeric vector with probability i-th hypothesis is an alternative}

\item{alphas}{Numeric vector of first parameter of alternative Beta distribution (Beta(a_i, 1)).}

\item{numerator_bh}{If TRUE (default is FALSE) then replace pi0 in numerator by 1.}
}
\value{
Numeric with Tail fdr value.
}
\description{
Tail fdr for beta-mixture model
}
