% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamix.R
\name{get_thresholds_betamix}
\alias{get_thresholds_betamix}
\title{Thresholds for beta-mixture model}
\usage{
get_thresholds_betamix(c, pi1s, alphas)
}
\arguments{
\item{c}{Numeric constant, the contour level ofthe local fdrs}

\item{pi1s}{Numeric vector with probability i-th hypothesis is an alternative}

\item{alphas}{Numeric vector of first parameter of alternative Beta distribution (Beta(a_i, 1)).}
}
\value{
Vector of thresholds t_i so that lfdr_i(t_i) = c
}
\description{
Thresholds for beta-mixture model
}
