% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_simulations_functions.R
\name{grouped_global_null_sim}
\alias{grouped_global_null_sim}
\title{Simulation: Grouped multiple testing under global null}
\usage{
grouped_global_null_sim(m, K)
}
\arguments{
\item{m}{Number of hypotheses}

\item{K}{Number of groups}
}
\value{
Data frame with columns `Hs` (null or alternative), `Ps` (p-value) and `Xs` (side-information)
}
\description{
Simulation: Grouped multiple testing under global null
}
