% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_simulations_functions.R
\name{grouped_sim}
\alias{grouped_sim}
\title{Simulation: Grouped multiple testing  with varying pi_0}
\usage{
grouped_sim(m, K_coarse, pi0_global, sparsity_multiplier = 4, K = 40)
}
\arguments{
\item{m}{Number of hypotheses}

\item{pi0_global}{Overall proportion of null hypotheses}

\item{sparsity_multiplier}{For each group with alternatives, there are `sparsity_multiplier`-1 groups with alternatives.}

\item{K}{Number of latent groups}
}
\value{
Data frame with columns `Hs` (null or alternative), `Ps` (p-value) and `Xs` (side-information)
}
\description{
Simulation: Grouped multiple testing  with varying pi_0
}
