% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{grouped_storey_weighter}
\alias{grouped_storey_weighter}
\title{Learn multiple testing weights based on the GBH procedure}
\usage{
grouped_storey_weighter(Ps, Xs, Xs_new, tau, alpha)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values in training set.}

\item{Xs}{Factor of groups in training set.}

\item{Xs_new}{Factor of groups in test-set (out-of-fold).}

\item{tau}{Numeric (default = 0.5), the level at which tau-censoring is applied.}

\item{alpha}{Significance level at which to apply method (not used for this weight learning method)}
}
\value{
Numeric vector with weights of test-set hypotheses
}
\description{
Learn multiple testing weights based on the GBH procedure
}
