% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwise_sabha_pi0.R
\name{groupwise_sabha}
\alias{groupwise_sabha}
\title{SABHA multiple testing procedure for grouped side-information}
\usage{
groupwise_sabha(
  pvals,
  Xs,
  alpha,
  tau = 0.5,
  eps = 0.1,
  return_fit = FALSE,
  beta = 10^3,
  max_iters = 400
)
}
\arguments{
\item{pvals}{Vector of P-values.}

\item{Xs}{Vector with categorical side-information}

\item{alpha}{Nominal testing level}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
The code has been modified from https://www.stat.uchicago.edu/~rina/sabha.html to be faster in situations
with large m and many groups.
}
