% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{ihw_betamix_censored}
\alias{ihw_betamix_censored}
\title{The IHW-BH/Storey-Betamix multiple testing procedure (with censoring)}
\usage{
ihw_betamix_censored(
  Ps,
  Xs,
  alpha,
  tau = 0.1,
  Storey = FALSE,
  numerator_bh = TRUE,
  ...
)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Xs}{Numeric vector or matrix with covariates}

\item{alpha}{Significance level at which to apply method}

\item{tau}{Numeric (default = 0.1), the level at which tau-censoring is applied.}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}

\item{numerator_bh}{Bool (default:TRUE) Define the numerator of the marginal FDR as sum pi_0(X_i) t(X_i), if FALSE, or sum t(X_i) if TRUE}

\item{...}{Other arguments passed to `censored_betamix_weighter`}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
The IHW-BH/Storey-Betamix multiple testing procedure (with censoring)
}
