% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{ihw_bh}
\alias{ihw_bh}
\title{Main wrapper for general IHW-BH/Storey procedure}
\usage{
ihw_bh(
  primary_stat,
  Xs,
  alpha,
  wt_fitter,
  tau = 0.5,
  folds = NULL,
  kfolds = 5L,
  Storey = FALSE,
  stat_type = "pvalue",
  return_weights = FALSE,
  ...
)
}
\arguments{
\item{primary_stat}{Vector of test statistics, typically pvalues (see `stat_type`)}

\item{Xs}{Side-information for each test}

\item{alpha}{Significance level at which to apply method}

\item{wt_fitter}{Function that does the weight learning, for example `grouped_storey_weighter`.}

\item{tau}{Numeric (default: 0.5) censoring level}

\item{folds}{Factor with assignments of tests to folds, defaults to NULL in which case they are chosen at random.}

\item{kfolds}{Integer, number of folds into which to split hypotheses (ignored if `folds` is not NULL).}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}

\item{stat_type}{Either "pvalue" (default) or "zscore"}

\item{return_weights}{Boolean, whether to return weigts (defauls to FALSE)}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Main wrapper for general IHW-BH/Storey procedure
}
