% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{ihw_cars}
\alias{ihw_cars}
\title{The IHW-BH/Storey-CARS multiple testing procedure}
\usage{
ihw_cars(primary_stats, Xs, alpha, Storey = FALSE, ...)
}
\arguments{
\item{primary_stats}{Numeric vector of z-score statistics (converted to p-values via 2*(1-pnorm(abs(primary_stats)))).}

\item{Xs}{Numeric vector or matrix with covariates}

\item{alpha}{Significance level at which to apply method}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}

\item{...}{Other arguments passed to `cars_weighter`}

\item{tau}{Numeric (default: 0.5) level at which to apply Storey's pi0 estimator}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
The IHW-BH/Storey-CARS multiple testing procedure
}
