% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{ihw_gbh}
\alias{ihw_gbh}
\title{The IHW-GBH multiple testing procedure}
\usage{
ihw_gbh(Ps, Xs, alpha, tau = 0.5, Storey = FALSE, ...)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Xs}{Factor to which different hypotheses belong}

\item{alpha}{Significance level at which to apply method}

\item{tau}{Numeric (default: 0.5) level at which to apply Storey's pi0 estimator}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
The IHW-GBH multiple testing procedure
}
