% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{ihw_nmeth_wrapper}
\alias{ihw_nmeth_wrapper}
\title{Wrapper for the IHW procedure}
\usage{
ihw_nmeth_wrapper(Ps, Xs, alpha, pre_bin = FALSE, Storey = FALSE)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Xs}{Vector or matrix of covariates}

\item{alpha}{Significance level at which to apply method}

\item{pre_bin}{Keyword argument that defines the binning procedure}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Wrapper for the IHW procedure
}
