% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_fdr.R
\name{oracle_local_fdr_test}
\alias{oracle_local_fdr_test}
\title{FDR control procedure based on local FDRs}
\usage{
oracle_local_fdr_test(unadj_p, lfdrs, alpha)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values. This is only used to break ties.}

\item{lfdrs}{Numeric vector with Local false discovery rate of each hypothesis}

\item{alpha}{Significance level at which to apply method}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
FDR control procedure based on local FDRs
}
