% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_methods.R
\name{stratified_bhq}
\alias{stratified_bhq}
\title{stratified_bhq: Stratified Benjamini Hochberg}
\usage{
stratified_bhq(unadj_p, groups, alpha)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{groups}{Factor to which different hypotheses belong}

\item{alpha}{Significance level at which to apply method}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
stratified_bhq: Stratified Benjamini Hochberg
}
\references{
Sun, Lei, et al. "Stratified false discovery control for large-scale hypothesis testing with application to genome-wide
   association studies." Genetic epidemiology 30.6 (2006): 519-530.

Also see similar function in IHWpaper package on Bioconductor.
}
