% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_methods.R
\name{stratified_clfdr}
\alias{stratified_clfdr}
\title{stratified clfdr: local fdr based method}
\usage{
stratified_clfdr(unadj_p, groups, alpha)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{groups}{Factor to which different hypotheses belong}

\item{alpha}{Significance level at which to apply method}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
The two-groups model is estimated using fdrtool.
}
\references{
Cai, T. Tony, and Wenguang Sun. "Simultaneous testing of grouped hypotheses: Finding needles in multiple haystacks."
          Journal of the American Statistical Association 104.488 (2009).

Also see similar function in IHWpaper package on Bioconductor.
}
