% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_and_weighting.R
\name{tau_weighted_bh}
\alias{tau_weighted_bh}
\title{The tau-weighted BH multiple testing procedure}
\usage{
tau_weighted_bh(Ps, ws, tau = 0.5)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{ws}{Numeric vector of multiple testing weights}

\item{tau}{Numeric (default = 0.5), the level at which tau-censoring is applied.}
}
\value{
Vector of adjusted p-values
}
\description{
The tau-weighted BH multiple testing procedure
}
