% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamix.R
\name{weights_betamix}
\alias{weights_betamix}
\title{Optimal weights based on a fitted conditional Beta-mixture model}
\usage{
weights_betamix(alpha, pi1s, alphas, numerator_bh = FALSE)
}
\arguments{
\item{alpha}{Nominal testing level}

\item{pi1s}{Numeric vector with probability i-th hypothesis is an alternative}

\item{alphas}{Numeric vector of first parameter of alternative Beta distribution (Beta(a_i, 1)).}

\item{numerator_bh}{If TRUE (default is FALSE) then replace pi0 in numerator by 1.}
}
\value{
Numeric vector with weights.
}
\description{
Optimal weights based on a fitted conditional Beta-mixture model
}
